Cleanup is only triggered if there are empty directories.

+++ photosort.sh
@@ -83,6 +83,8 @@ exiftool -P -d "$ARCHIVE/%Y/%m/%Y%m%d_%H%M%S" -ext mov -ext jpg \
 # so that people may add entire folders, as this eases the whole
 # copy process.
 EMPTY_DIRS=$(find $MONITOR -type d -empty)
-rmdir -p $EMPTY_DIRS
+if [ ! -z ${EMPTY_DIRS} ]; then
+    rmdir -p $EMPTY_DIRS
+fi
 
 rm $LOCK
